requirements.txt
flask==2.0.1
google-generativeai>=0.1.0
selenium==4.0.0
healenium-client==2.1.0
allure-pytest==2.9.43
pytest==6.2.5
python-owasp-zap==0.1.0
lighthouse-parser==0.3.0